Move integration tests into separate package
