+++ monthly-contributions/src/test/scala/com/gu/support/workers/integration/CreateZuoraSubscriptionSpec.scala
@@ -1,4 +1,4 @@
-package com.gu.support.workers.lambdas
+package com.gu.support.workers.integration
 
 import java.io.ByteArrayOutputStream
 
@@ -6,6 +6,7 @@ import com.gu.support.workers.Conversions.{FromOutputStream, StringInputStreamCo
 import com.gu.support.workers.Fixtures._
 import com.gu.support.workers.LambdaSpec
 import com.gu.support.workers.encoding.StateCodecs._
+import com.gu.support.workers.lambdas.CreateZuoraSubscription
 import com.gu.support.workers.model.monthlyContributions.state.SendThankYouEmailState
 import com.gu.test.tags.annotations.IntegrationTest
