+++ monthly-contributions/src/test/scala/com/gu/support/workers/integration/SendThankYouEmailSpec.scala
@@ -1,10 +1,11 @@
-package com.gu.support.workers.lambdas
+package com.gu.support.workers.integration
 
 import java.io.ByteArrayOutputStream
 
 import com.gu.support.workers.Conversions.{FromOutputStream, StringInputStreamConversions}
 import com.gu.support.workers.Fixtures.thankYouEmailJson
 import com.gu.support.workers.LambdaSpec
+import com.gu.support.workers.lambdas.SendThankYouEmail
 import com.gu.test.tags.annotations.IntegrationTest
 
 @IntegrationTest
